issue #112 rename : AddFlightRepository FlightRepository
